+++ Plugins/Vorlon/plugins/modernizrReport/vorlon.modernizrReport.interfaces.ts
@@ -1,0 +1,7 @@
+﻿module VORLON {
+    export class FeatureSupported {
+        public featureName: string;
+        public isSupported: boolean;
+        public type: string;
+    }
+}