+++ Server/Scripts/typings/vorlon/vorlon.dashboardPlugin.d.ts
@@ -1,0 +1,20 @@
+declare module VORLON {
+    class DashboardPlugin extends BasePlugin {
+        htmlFragmentUrl: any;
+        cssStyleSheetUrl: any;
+        DashboardCommands: any;
+        constructor(name: string, htmlFragmentUrl: string, cssStyleSheetUrl: string);
+        Type: PluginType;
+        debug: boolean;
+        getID(): string;
+        isReady(): boolean;
+        startDashboardSide(div: HTMLDivElement): void;
+        onRealtimeMessageReceivedFromClientSide(receivedObject: any): void;
+        sendToClient(data: any): void;
+        sendCommandToClient(command: string, data?: any, incrementVisualIndicator?: boolean): void;
+        sendCommandToPluginClient(pluginId: string, command: string, data?: any, incrementVisualIndicator?: boolean): void;
+        sendCommandToPluginDashboard(pluginId: string, command: string, data?: any, incrementVisualIndicator?: boolean): void;
+        _insertHtmlContentAsync(divContainer: HTMLDivElement, callback: (filledDiv: HTMLDivElement) => void): void;
+        private _stripContent(content);
+    }
+}